Create os_pinger.py
Total python noob here. I tried to create a pinger as well and this is what I came up with. I don't have any I/O, but I like the part where it shows the best n worlds you might consider. Not sure if this helps, but if it gives you any ideas, feel free to use it in any way.